SPARC port: shim/stdafx/doc.h fills — KEEP+SHIM passes 5→7
Overnight infra build-out toward the 13-file doc.h-includes wall.

port/mfc_shim.h additions:
  * CTypedPtrMap<BASE,KEY,VALUE> 3-arg template (xml/xmlparse.h's
    CAttributeMap, plus various string→ptr maps in OtherTypes.h)
  * CTypedPtrArray<BASE,T> + CPtrArray (CAliasArray / CTriggerArray
    / CVariableArray in OtherTypes.h)
  * tStringToStringMap / tStringMapOfMaps typedefs (ported from
    stdafx.h.windows-original line 158-159; used by plugins.h,
    doc.h, scripting/methods/methods_arrays.cpp)
  * BLACK..WHITE enum (ported from stdafx.h.windows-original line 191)
  * CTime int-assignment + int-construction so `tWhenMatched = 0`
    and CmcDateTime epoch initialisations compile
  * POSITION-walk methods on CMapBase (GetStartPosition, GetNextAssoc)
  * COM scalar stubs (DISPID, IDispatch, BSTR, VARIANT, EXCEPINFO,
    REFIID, GUID, ITypeInfo, __int64) — let plugins.h compile
    without dragging the WSH path
  * Win32 NLS helpers: SUBLANG_*, LANG_NEUTRAL, lstrcpyn

port/stdafx.h additions (mostly migrated from stdafx.h.windows-original):
  * <climits>, <cfloat>, <zlib.h>, regexp.h, format.h pulled in
    early so doc.h's z_stream / Bytef / t_regexp / CFormat members
    compile.
  * IS_SET / SET_BIT / REMOVE_BIT / TOGGLE_BIT bit-mask macros
  * CHECK_BIT_NUMBER / SET_BIT_NUMBER / CLEAR_BIT_NUMBER /
    TOGGLE_BIT_NUMBER bit-number macros
  * DELETE_LIST(listname) and DELETE_MAP(mapname,pointertype) —
    drain helpers used in CXMLelement::~CXMLelement and elsewhere
  * MAX_CUSTOM, OTHER_CUSTOM, SAMECOLOUR, NO_COLOUR, NUMITEMS,
    PLURAL/PLURALIE/PLURALES, DEFAULT_CHAT_PORT, NOSOUNDLIT
  * eKeypad_0..eCtrl_Keypad_Plus enum + eKeypad_Max_Items
    (verbatim from stdafx.h.windows-original line 588)
  * POSITION typedef (void *), CStringList class, CEvent stub,
    CChatSocketList typedef, CProgressDlg forward decl,
    full CFindInfo class definition (ported from windows-original)
  * sqlite3 / sqlite3_stmt forward decls (real impl plugs in later)
  * std::binary_function unqualified-name pull (still ships in
    libstdc++ 15 with -std=c++17, just not auto-imported)
  * PLUGIN_UNIQUE_ID_LENGTH

doc.h:
  * Wrap GUI-laden include block (worldsock.h, chatsock.h,
    chatlistensock.h, dialogs/*, TimerWnd.h, paneline.h,
    miniwindow.h, plus xml/xmlparse + plugins + version) with
    `#ifdef __SPARC_SOLARIS7_PORT__` → forward-decls + the
    minimal real includes (xml/xmlparse, OtherTypes, plugins,
    version, plus regexp + mcdatetime). The non-port path is
    unchanged.

port/wx/main.cpp:
  * Replace %zu with %lu + cast — Solaris 7 printf is C89 and
    prints "%zu" literally as the characters "zu". Fix in two
    INIT_ARRAY walker fprintf calls.

KEEP+SHIM compile tally now: PASS 7/20 (was 5/18):
  + stdafx.cpp (new)
  + exceptions.cpp (new)
  + Color.cpp Dmetaph.cpp Replace.cpp format.cpp mcdatetime.cpp

Remaining 13 still hit doc.h's deeper deps (deque<>, CView,
CFrameWnd, CFindInfo nested usages, MiniWindowMap, HACCEL,
LPDIRECTSOUNDBUFFER, CPaneStyle, accelerators, splitter,
sendvw types). Each round of stubbing surfaces 5+ new symbols;
the proper structural fix is the doc_core.h split called out in
PORT_STATUS.md, not endless ifdefs.

+++ port/wx/main.cpp
@@ -271,9 +271,11 @@ extern "C" {
         }
         if (!init_array || !init_array_sz) return 0;
 
-        std::fprintf(stderr, "[sst-init] %s: walking %zu init_array entries\n",
+        // Solaris 7 printf doesn't grok %zu — prints literal "zu".
+        // Cast to unsigned long and use %lu to match.
+        std::fprintf(stderr, "[sst-init] %s: walking %lu init_array entries\n",
                      info->dlpi_name && *info->dlpi_name ? info->dlpi_name : "(self)",
-                     init_array_sz);
+                     static_cast<unsigned long>(init_array_sz));
         for (std::size_t i = 0; i < init_array_sz; ++i) {
             if (init_array[i]) init_array[i](0, nullptr, nullptr);
         }
@@ -293,7 +295,8 @@ extern "C" {
         // for our wxApp etc. live here, not in any DSO).
         if (__init_array_start && __init_array_end) {
             std::size_t n = __init_array_end - __init_array_start;
-            std::fprintf(stderr, "[sst-init] (main exe): walking %zu init_array entries\n", n);
+            std::fprintf(stderr, "[sst-init] (main exe): walking %lu init_array entries\n",
+                         static_cast<unsigned long>(n));
             for (void (**fn)(int, char **, char **) = __init_array_start;
                  fn < __init_array_end; ++fn)
             {
